stage4/projects/p02-yolov11-realtime/pytorch/requirements.txt
# YOLOv11 PyTorch Requirements
# Python >=3.8

# Core
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Vision
opencv-python>=4.8.0
pillow>=10.0.0

# Data
numpy>=1.23.0
pyyaml>=6.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Metrics
scikit-learn>=1.3.0
pandas>=2.0.0

# Export (optional)
onnx>=1.14.0
onnxruntime>=1.15.0
# tensorrt>=8.6.0  # NVIDIA GPU only

# Development (optional)
# pytest>=7.4.0
# black>=23.7.0
# flake8>=6.1.0
